Initial attempt at updated spidemonkey bindings

+++ tealeaf/platform/OpenGLView.mm
@@ -114,7 +114,7 @@ static volatile BOOL m_ogl_in = NO; // In OpenGL calls right now?
 	
 	if (m_ogl_en) {
 		CFTimeInterval timestamp = CFAbsoluteTimeGetCurrent();
-		int dt = (int)(1000 * (timestamp - last_timestamp));
+		long dt = (long)(1000 * (timestamp - last_timestamp));
 		core_tick(dt);
 		
 		// Measure time to perform a tick
